Enforce services to be closures
As was discussed in issue #8, having any *callable* value be a service leads
to issues, because function names are callable. Thus any parameter value
that is a global function name would have to be protected.

The easiest solution is to only treat closures as services, and treat
everything else as a parameter. This is actually the expected behavior,
according to the documentation.