Introduce static versions of share() and protect()
I have a use case where I want to be able to pass parameters to an
object that will then at some point create and initialize a pimple
container. Then set the parameters. I want to be able to override
defaults, including services.

Essentially, this:

    createContainer([
        foo => Pimple::shareStatic(function () { return new Foo(); }),
    ]);

Since the container does not exist yet, I have no instance to call
share() on. Therefore, I want static versions of share and protect,
so that this use case is supported properly.