Awareness capabilities to instances created by Pimple
If a class either implements the interface or uses the trait
will automatically be able to call the depency injection container
in its internal scope.

This can be very usefull when sometimes, you need access services from
an instance in the container, but is not convenient send the instance
of the service in the constructor (IoC).